move assert function to custom_assertions.py
